+++ Source/Atomic/Network/HttpRequest.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2008-2014 the Urho3D project.
+// Copyright (c) 2008-2015 the Urho3D project.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,11 +20,11 @@
 // THE SOFTWARE.
 //
 
-#include "Precompiled.h"
-#include "../Network/HttpRequest.h"
-#include "../IO/Log.h"
+#include "../Precompiled.h"
+
 #include "../Core/Profiler.h"
-#include "../Core/Timer.h"
+#include "../IO/Log.h"
+#include "../Network/HttpRequest.h"
 
 #include <Civetweb/include/civetweb.h>
 
@@ -50,9 +50,9 @@ HttpRequest::HttpRequest(const String& url, const String& verb, const Vector<Str
     // Size of response is unknown, so just set maximum value. The position will also be changed
     // to maximum value once the request is done, signaling end for Deserializer::IsEof().
     size_ = M_MAX_UNSIGNED;
-    
+
     LOGDEBUG("HTTP " + verb_ + " request to URL " + url_);
-    
+
     // Start the worker thread to actually create the connection and read the response data.
     Run();
 }
@@ -68,7 +68,7 @@ void HttpRequest::ThreadFunction()
     String host;
     String path = "/";
     int port = 80;
-    
+
     unsigned protocolEnd = url_.Find("://");
     if (protocolEnd != String::NPOS)
     {
@@ -77,24 +77,24 @@ void HttpRequest::ThreadFunction()
     }
     else
         host = url_;
-    
+
     unsigned pathStart = host.Find('/');
     if (pathStart != String::NPOS)
     {
         path = host.Substring(pathStart);
         host = host.Substring(0, pathStart);
     }
-    
+
     unsigned portStart = host.Find(':');
     if (portStart != String::NPOS)
     {
         port = ToInt(host.Substring(portStart + 1));
         host = host.Substring(0, portStart);
     }
-    
+
     char errorBuffer[ERROR_BUFFER_SIZE];
     memset(errorBuffer, 0, sizeof(errorBuffer));
-    
+
     String headersStr;
     for (unsigned i = 0; i < headers_.Size(); ++i)
     {
@@ -103,7 +103,7 @@ void HttpRequest::ThreadFunction()
         if (header.Length())
             headersStr += header + "\r\n";
     }
-    
+
     // Initiate the connection. This may block due to DNS query
     /// \todo SSL mode will not actually work unless Civetweb's SSL mode is initialized with an external SSL DLL
     mg_connection* connection = 0;
@@ -125,11 +125,11 @@ void HttpRequest::ThreadFunction()
             "\r\n"
             "%s", verb_.CString(), path.CString(), host.CString(), headersStr.CString(), postData_.Length(), postData_.CString());
     }
-    
+
     {
         MutexLock lock(mutex_);
         state_ = connection ? HTTP_OPEN : HTTP_ERROR;
-        
+
         // If no connection could be made, store the error and exit
         if (state_ == HTTP_ERROR)
         {
@@ -137,7 +137,7 @@ void HttpRequest::ThreadFunction()
             return;
         }
     }
-    
+
     // Loop while should run, read data from the connection, copy to the main thread buffer if there is space
     while (shouldRun_)
     {
@@ -145,29 +145,29 @@ void HttpRequest::ThreadFunction()
         int bytesRead = mg_read(connection, httpReadBuffer_.Get(), READ_BUFFER_SIZE / 4);
         if (bytesRead <= 0)
             break;
-        
+
         mutex_.Acquire();
-        
+
         // Wait until enough space in the main thread's ring buffer
         for (;;)
         {
             unsigned spaceInBuffer = READ_BUFFER_SIZE - ((writePosition_ - readPosition_) & (READ_BUFFER_SIZE - 1));
             if ((int)spaceInBuffer > bytesRead || !shouldRun_)
                 break;
-            
+
             mutex_.Release();
             Time::Sleep(5);
             mutex_.Acquire();
         }
-        
+
         if (!shouldRun_)
         {
             mutex_.Release();
             break;
         }
-        
+
         if (writePosition_ + bytesRead <= READ_BUFFER_SIZE)
-            memcpy(readBuffer_.Get() + writePosition_, httpReadBuffer_.Get(), bytesRead);
+            memcpy(readBuffer_.Get() + writePosition_, httpReadBuffer_.Get(), (size_t)bytesRead);
         else
         {
             // Handle ring buffer wrap
@@ -176,16 +176,16 @@ void HttpRequest::ThreadFunction()
             memcpy(readBuffer_.Get() + writePosition_, httpReadBuffer_.Get(), part1);
             memcpy(readBuffer_.Get(), httpReadBuffer_.Get() + part1, part2);
         }
-        
+
         writePosition_ += bytesRead;
         writePosition_ &= READ_BUFFER_SIZE - 1;
-        
+
         mutex_.Release();
     }
-    
+
     // Close the connection
     mg_close_connection(connection);
-    
+
     {
         MutexLock lock(mutex_);
         state_ = HTTP_CLOSED;
@@ -195,7 +195,7 @@ void HttpRequest::ThreadFunction()
 unsigned HttpRequest::Read(void* dest, unsigned size)
 {
     mutex_.Acquire();
-    
+
     unsigned char* destPtr = (unsigned char*)dest;
     unsigned sizeLeft = size;
     unsigned totalRead = 0;
@@ -203,7 +203,7 @@ unsigned HttpRequest::Read(void* dest, unsigned size)
     for (;;)
     {
         unsigned bytesAvailable;
-        
+
         for (;;)
         {
             bytesAvailable = CheckEofAndAvailableSize();
@@ -214,12 +214,12 @@ unsigned HttpRequest::Read(void* dest, unsigned size)
             Time::Sleep(5);
             mutex_.Acquire();
         }
-        
+
         if (bytesAvailable)
         {
             if (bytesAvailable > sizeLeft)
                 bytesAvailable = sizeLeft;
-            
+
             if (readPosition_ + bytesAvailable <= READ_BUFFER_SIZE)
                 memcpy(destPtr, readBuffer_.Get() + readPosition_, bytesAvailable);
             else
@@ -230,18 +230,18 @@ unsigned HttpRequest::Read(void* dest, unsigned size)
                 memcpy(destPtr, readBuffer_.Get() + readPosition_, part1);
                 memcpy(destPtr + part1, readBuffer_.Get(), part2);
             }
-            
+
             readPosition_ += bytesAvailable;
             readPosition_ &= READ_BUFFER_SIZE - 1;
             sizeLeft -= bytesAvailable;
             totalRead += bytesAvailable;
             destPtr += bytesAvailable;
         }
-        
+
         if (!sizeLeft || !bytesAvailable)
             break;
     }
-    
+
     // Check for end-of-file once more after reading the bytes
     CheckEofAndAvailableSize();
     mutex_.Release();
